+++ src/JMS/Serializer/Exclusion/DisjunctExclusionStrategy.php
@@ -1,0 +1,65 @@
+<?php
+
+namespace JMS\Serializer\Exclusion;
+
+use JMS\Serializer\Context;
+use JMS\Serializer\Metadata\ClassMetadata;
+use JMS\Serializer\Metadata\PropertyMetadata;
+
+/**
+ * Disjunct Exclusion Strategy.
+ *
+ * This strategy is short-circuiting and will skip a class, or property as soon as one of the delegates skips it.
+ *
+ * @author Johannes M. Schmitt <schmittjoh@gmail.com>
+ */
+class DisjunctExclusionStrategy implements ExclusionStrategyInterface
+{
+    private $delegates;
+
+    /**
+     * @param ExclusionStrategyInterface[] $delegates
+     */
+    public function __construct(array $delegates)
+    {
+        $this->delegates = $delegates;
+    }
+
+    /**
+     * Whether the class should be skipped.
+     *
+     * @param ClassMetadata $metadata
+     * @param Context $navigatorContext
+     *
+     * @return boolean
+     */
+    public function shouldSkipClass(ClassMetadata $metadata, Context $context)
+    {
+        foreach ($this->delegates as $delegate) {
+            if ($delegate->shouldSkipClass($metadata, $context)) {
+                return true;
+            }
+        }
+
+        return false;
+    }
+
+    /**
+     * Whether the property should be skipped.
+     *
+     * @param PropertyMetadata $property
+     * @param Context $navigatorContext
+     *
+     * @return boolean
+     */
+    public function shouldSkipProperty(PropertyMetadata $property, Context $context)
+    {
+        foreach ($this->delegates as $delegate) {
+            if ($delegate->shouldSkipProperty($property, $context)) {
+                return true;
+            }
+        }
+
+        return false;
+    }
+}